requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
supabase==2.16.0
PyJWT==2.8.0
google-generativeai==0.3.2
python-docx==1.1.0
PyPDF2==3.0.1
Pillow==11.0.0
requests==2.31.0
python-multipart==0.0.9
python-dotenv==1.0.1
filetype==1.2.0

